Original change: https://android-review.googlesource.com/c/platform/external/perfetto/+/2864529

+++ ui/src/tracks/visualised_args/index.ts
@@ -80,7 +80,9 @@ export class VisualisedArgsTrack extends ChromeSliceTrack {
   }
 
   async onDestroy(): Promise<void> {
-    this.engine.query(`drop view ${this.helperViewName}`);
+    if (this.engine.isAlive) {
+      this.engine.query(`drop view ${this.helperViewName}`);
+    }
   }
 
   getFont() {
